+++ .github/workflows/github_update_todo.py
@@ -1,4 +1,6 @@
 """Checks repository in main and prints the content for a ToDo issue."""
+from __future__ import annotations
+
 from pylint import epylint as lint