+++ puma/hlplots/__init__.py
@@ -1,4 +1,5 @@
 """High level plotting API within puma, to avoid code duplication."""
+from __future__ import annotations
 
 from puma.hlplots.results import Results
 from puma.hlplots.tagger import Tagger