+++ puma/tests/test_line_2d.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 """Unit test script for the functions in line_plot_2d.py."""
+from __future__ import annotations
 
 import os
 import tempfile
@@ -23,7 +24,7 @@ class Line2DTestCase(unittest.TestCase):
         with self.assertRaises(ValueError):
             Line2D(
                 x_values="Test",
-                y_values=int(5),
+                y_values=5,
             )
 
     def test_differnt_input_types(self):