+++ puma/tests/test_metrics.py
@@ -2,6 +2,7 @@
 
 
 """Unit test script for the functions in metrics.py."""
+from __future__ import annotations
 
 import unittest