+++ puma/tests/utils/test_generate.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 """Unit test script for the functions in utils/generate.py."""
+from __future__ import annotations
 
 import unittest