+++ puma/var_vs_eff.py
@@ -1,4 +1,6 @@
 """Efficiency plots vs. specific variable."""
+from __future__ import annotations
+
 import numpy as np
 
 # TODO: fix the import below
@@ -21,7 +23,7 @@ class VarVsEff(VarVsVar):  # pylint: disable=too-many-instance-attributes
         x_var_bkg: np.ndarray = None,
         disc_bkg: np.ndarray = None,
         bins=10,
-        working_point: float = None,
+        working_point: float | None = None,
         disc_cut=None,
         flat_per_bin: bool = False,
         key: str = None,